factorial and removeDupli

+++ src/com/test/Factorial.java
@@ -7,12 +7,12 @@ public class Factorial {
     public static void main(String[] args){
         Scanner s = new Scanner(System.in);
         System.out.println("Enter a number : ");
-        int n = s.nextInt();
+        double n = s.nextInt();
 
         System.out.println("Factorial : "+ factorial(n));
     }
 
-    public static long factorial(long n){
+    public static double factorial(double n){
         if(n == 1 )
             return 1;
         else
